give each node its id (fixing imgui tree id problems)

+++ hw3d/App.cpp
@@ -87,7 +87,7 @@ void App::DoFrame()
 	{
 		if( !wnd.CursorEnabled() )
 		{
-			cam.Rotate( delta->x,delta->y );
+			cam.Rotate( (float)delta->x,(float)delta->y );
 		}
 	}
